Move test-runner.ts to /query-testing/test-runner.ts
